logout endpoint tests pass

+++ server/src/main/java/service/LogoutService.java
@@ -16,6 +16,13 @@ public class LogoutService {
 
     public LogoutResult logout(LogoutRequest request) throws ResponseException {
         try {
+
+            var authToken = authDAO.getAuth(request.authToken());
+
+            if (authToken == null) {
+                throw new ResponseException(401, "Error: unauthorized");
+            }
+
             authDAO.deleteAuth(request.authToken());
 
             return new LogoutResult("logged out");
